Thêm chức năng hạ cấp seller thành bidder + update lại các tính năng xóa/xác nhận thánh seller

+++ controllers/usercontroller.js
@@ -568,27 +568,35 @@ class userController {
     }
     //----------post----------------------
     async setPostRegistConfirm(req, res) {
-        var name = req.params.id;
+        var name = req.body.user;
         var acc = {};
         await db.findOne({ name }).then(doc => {
             acc = doc;
         })
-        var myquery = {
-            _id: ObjectId(acc._id)
+        if(acc.length===0)
+        {
+            res.redirect('/users/manageuser/register');
+
         }
-        var changeAcc = {
-            status: "Seller",
-        };
-        var options = {
-            multi: true
+        else{
+            var myquery = {
+                _id: ObjectId(acc._id)
+            }
+            var changeAcc = {
+                status: "Seller",
+            };
+            var options = {
+                multi: true
+            }
+            // usermodels.UpdateInfoAccount(changeAcc,iduser);
+            await db.update(myquery, changeAcc, options);
+            await dbregisterseller.findOneAndRemove({ name });
+            res.redirect('/users/manageuser/register');
         }
-        // usermodels.UpdateInfoAccount(changeAcc,iduser);
-        await db.update(myquery, changeAcc, options);
-        await dbregisterseller.findOneAndRemove({ name });
-        res.redirect('/users/manageuser/register');
+        
     }
     async setPostRegistDelete(req, res) {
-        var name = req.params.id;
+        var name = req.body.id;
         await dbregisterseller.findOneAndRemove({ name });
         res.redirect('/users/manageuser/register');
     }
@@ -628,6 +636,33 @@ class userController {
         // usermodels.UpdateInfoAccount(changeAcc,iduser);
         await dbcategory.update(myquery, changeCate, options);
         res.redirect('/users/managecategory');
+    }
+    async setPostCancelSeller(req,res){
+        var name = req.body.leveldown;
+        var acc = {};
+        await db.findOne({ name }).then(doc => {
+            acc = doc;
+        })
+        if(acc.length===0)
+        {
+            res.redirect('/users/manageuser/register');
+
+        }
+        else{
+            var myquery = {
+                _id: ObjectId(acc._id)
+            }
+            var changeAcc = {
+                status: "Bidder",
+            };
+            var options = {
+                multi: true
+            }
+            // usermodels.UpdateInfoAccount(changeAcc,iduser);
+            await db.update(myquery, changeAcc, options);
+            await dbregisterseller.findOneAndRemove({ name });
+            res.redirect('/users/manageuser/register');
+        }
     }
 
 }
